Add cached OSM node and ways support, usd bigint fro OSM ID
Create poi_osm_cache table and add all ways, related nodes and nodes to this table during the processing.
It speed ups the processing also the XML export will contain not only the ways but the nodes of the way. This was required by JOSM processing.
One missing piece is currently the ways' nodes will lose all the specified tags. This is not a big problem so far, but will be fixed soon.