Put POIs into a building with same address
Looking for a building (actually a way) around the POI node with same address with within a preconfigured distance.

Actually this method helps to find a building for a single node.

Currently the node moved to the center of the building. Later we have to improve the method of the POI positioning.

Fixes #34