Move OSM database import's projection to WGS84 latlong (EPSG:4326) from webmercator (EPSG:3857)
* Change the import setting of OSM data (osm2pgsql)
* Rewrite queries

Currently all data in OSM extract is loaded using webmercator (EPSG:3857) projection to the PostgreSQL/PostGIS database. This is the default for osm2pgsql and osm2pgsql -m parameter explicitly defines it during the import. It is optimal to generate tiles from it, however we use it to compare with available WGS84 lat, long positions in EPSG:4326 supplied by data providers. However this is not good for the performance, since we have to transform every query, and the DB will not able to use geom indexes at all. Change the default projection to WGS84 latlong (EPSG:4326) and check if it makes the life easier.

Issue #60